[chore]: refactor bootstrapping code
