+++ desktop-app/public/index.html
@@ -16,7 +16,7 @@
     <title>React App</title>
   </head>
   <body>
-    <div id="root"></div>
+    <div id="bootstrap"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
