+++ desktop-app/src/client-common/core/app/index.js
@@ -1,7 +1,8 @@
 // escherpad has all these insertion points
 const NAME_SPACE = "example-plugin";
+
 // this is almost DI.
-export const module = {
+const module = {
     requireReload: true,
     commands: {},
     reducer: {}, // or a function (s, a)=>s
@@ -10,3 +11,4 @@ export const module = {
     saga: {},
     config: (c) => c
 };
+export default module
